pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tradfibot"
version = "0.1.0"
description = "RL-based automated crypto trading engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.11.0",
    "ccxt>=4.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]
